+++ src/test/kotlin/dev/tailwindrainbow/intellij/application/highlight/TailwindClassParserTest.kt
@@ -26,7 +26,6 @@ class TailwindClassParserTest {
                     arbitrary = arbitraryStyle,
                     important = importantStyle,
                 ),
-                ignoredPrefixModifiers = setOf("group", "peer"),
             ),
         )