+++ src/test/kotlin/dev/tailwindrainbow/intellij/application/settings/SettingsFormMapperTest.kt
@@ -203,7 +203,6 @@ class SettingsFormMapperTest {
         classIdentifiers = "class",
         classFunctions = classFunctions,
         templateTags = "tw",
-        ignoredPrefixModifiers = "group",
         supportedExtensions = supportedExtensions,
         readsClassLikeStrings = true,
     )